subtask3 fix
